+++ views/index.ejs
@@ -1,5 +1,5 @@
 <h1>SurveyMe</h1>
 
-<button type="submit"><a href="./auth/login"> Login </a></button>
+<a href="./auth/login"><button type="submit"> Login </button></a> 
 <br>
-<button type="submit"><a href="./auth/register"> Register </a></button>
+<a href="./auth/register"><button type="submit"> Register </button></a>
