+++ views/user/profile.ejs
@@ -1,8 +1,5 @@
 <h2>Profile Page</h2>
-Asheela Magwili
 
 <% // Displays each user email and name %> 
-<% users.forEach(user => { %>
     <div> <%= user.name  %> </div>
     <div> <%= user.email  %> </div>
-<% }) %>
